feat(): migrate const && format some value
